Update video_player.py
I dont know open cv but i added some tkinter in your python script and made it like an app. But there is only one error. If i drag the video window the sound continues but video stucks. Because of that sound and video doesn't play at the same time. I cant solve that problem. So if you can, please make another update for that.